Refactor. Int for distance instead of Double.

+++ RunPersonalRecord/CoreData/ActivityEntity+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  ActivityEntity+CoreDataProperties.swift
 //  RunPersonalRecord
 //
-//  Created by Artsem Lemiasheuski on 20.06.20.
+//  Created by Artsem Lemiasheuski on 26.06.20.
 //  Copyright © 2020 metaxa.RunPersonalRecord. All rights reserved.
 //
 //
@@ -20,7 +20,7 @@ extension ActivityEntity {
     @NSManaged public var activityAttribute: Activity?
     @NSManaged public var completed: Bool
     @NSManaged public var date: Date?
-    @NSManaged public var distance: Double
+    @NSManaged public var distance: Int32
     @NSManaged public var duration: Double
     @NSManaged public var pace: Pace?
